+++ src/galaxy/Sector.cpp
@@ -43,7 +43,7 @@ void Sector::GetCustomSystems()
 //////////////////////// Sector
 Sector::Sector(int x, int y, int z)
 {
-	unsigned long _init[4] = { Uint32(x), Uint32(y), Uint32(z), UNIVERSE_SEED };
+	Uint32 _init[4] = { Uint32(x), Uint32(y), Uint32(z), UNIVERSE_SEED };
 	Random rng(_init, 4);
 
 	sx = x; sy = y; sz = z;
